# nextlinux_engine/services/apiext/swagger/swagger.yaml
swagger: "2.0"
info:
  description: "This is the Anchore Engine API. Provides the primary external API for users of the service."
  version: "0.1.20"
  title: "Anchore Engine API Server"
  contact:
    email: "nurmi@nextlinux.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
#externalDocs:
#  description: "Anchore Engine"
#  url: "http://github.com/nextlinux/nextlinux-engine"
...
basePath: "/"
schemes:
  - "http"
  - "https"
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /status:
    get:
      tags:
        - System
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.system
      operationId: get_status
      x-nextlinux-authz-action: None
      summary: Service status
      description: Get the API service status
      responses:
        200:
          description: Status listing
          schema:
            $ref: "#/definitions/StatusResponse"
        500:
          description: Internal error
          schema:
            $ref: "#/definitions/ApiErrorResponse"
...
  /imports/images:
    post:
      tags:
        - Imports
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.image_imports
      operationId: create_operation
      x-nextlinux-authz-action: importImage
      summary: "Begin the import of an image analyzed by Syft into the system"
      produces:
        - "application/json"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportOperation"
        500:
          description: "Internal Error"
    get:
      tags:
        - Imports
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.image_imports
      operationId: list_operations
      x-nextlinux-authz-action: importImage
      summary: "Lists in-progress imports"
      produces:
        - "application/json"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImports"
        500:
          description: "Internal Error"
  /imports/images/{operation_id}:
    get:
      tags:
        - Imports
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.image_imports
      operationId: get_operation
      x-nextlinux-authz-action: importImage
      summary: "Get detail on a single import"
      produces:
        - "application/json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportOperation"
        500:
          description: "Internal Error"
    delete:
      tags:
        - Imports
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.image_imports
      operationId: invalidate_operation
      x-nextlinux-authz-action: importImage
      summary: "Invalidate operation ID so it can be garbage collected"
      produces:
        - "application/json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportOperation"
        500:
          description: "Internal Error"
  /imports/images/{operation_id}/packages:
    get:
      tags:
        - Imports
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.image_imports
      operationId: list_import_packages
      x-nextlinux-authz-action: importImage
      summary: "List uploaded package manifests"
      produces:
        - "application/json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImportContentDigestList"
        500:
          description: "Internal Error"
    post:
      tags:
        - Imports
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.image_imports
      operationId: import_image_packages
      x-nextlinux-authz-action: importImage
      summary: "Begin the import of an image analyzed by Syft into the system"
      produces:
        - "application/json"
      consumes:
        - "application/json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
        - name: sbom
          in: body
          required: true
          schema:
            $ref: "#/definitions/ImagePackageManifest"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportContentResponse"
        500:
          description: "Internal Error"
  /imports/images/{operation_id}/dockerfile:
    get:
      tags:
        - Imports
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.image_imports
      operationId: list_import_dockerfiles
      x-nextlinux-authz-action: importImage
      summary: "List uploaded dockerfiles"
      produces:
        - "application/json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImportContentDigestList"
        500:
          description: "Internal Error"
    post:
      tags:
        - Imports
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.image_imports
      operationId: import_image_dockerfile
      x-nextlinux-authz-action: importImage
      summary: "Begin the import of an image analyzed by Syft into the system"
      produces:
        - "application/json"
      consumes:
        - "text/plain; utf-8"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
        - in: body
          name: contents
          required: true
          schema:
            type: string
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportContentResponse"
        500:
          description: "Internal Error"
  /imports/images/{operation_id}/manifest:
    get:
      tags:
        - Imports
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.image_imports
      operationId: list_import_image_manifests
      x-nextlinux-authz-action: importImage
      summary: "List uploaded image manifests"
      produces:
        - "application/json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImportContentDigestList"
        500:
          description: "Internal Error"
    post:
      tags:
        - Imports
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.image_imports
      operationId: import_image_manifest
      x-nextlinux-authz-action: importImage
      summary: "Import a docker or OCI distribution manifest to associate with the image"
      produces:
        - "application/json"
      consumes:
        - "application/vnd.oci.image.manifest.v1+json"
        - "application/vnd.docker.distribution.manifest.v2+json"
        - "application/vnd.docker.distribution.manifest.v1+json" # Older Docker spec
        - "application/vnd.docker.distribution.manifest.v1+prettyjws" # Signed version
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
        - in: body
          name: contents
          required: true
          schema:
            type: object
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportContentResponse"
        500:
          description: "Internal Error"
  /imports/images/{operation_id}/parent_manifest:
    get:
      tags:
        - Imports
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.image_imports
      operationId: list_import_parent_manifests
      x-nextlinux-authz-action: importImage
      summary: "List uploaded parent manifests (manifest lists for a tag)"
      produces:
        - "application/json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImportContentDigestList"
        500:
          description: "Internal Error"
    post:
      tags:
        - Imports
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.image_imports
      operationId: import_image_parent_manifest
      x-nextlinux-authz-action: importImage
      summary: "Import a docker or OCI distribution manifest list to associate with the image"
      produces:
        - "application/json"
      consumes:
        - "application/vnd.docker.distribution.manifest.list.v2+json"
        - "application/vnd.oci.image.index.v1+json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
        - in: body
          name: contents
          required: true
          schema:
            type: object
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportContentResponse"
        500:
          description: "Internal Error"
  /imports/images/{operation_id}/image_config:
    get:
      tags:
        - Imports
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.image_imports
      operationId: list_import_image_configs
      x-nextlinux-authz-action: importImage
      summary: "List uploaded image configs"
      produces:
        - "application/json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImportContentDigestList"
        500:
          description: "Internal Error"
    post:
      tags:
        - Imports
      x-swagger-router-controller: nextlinux_engine.services.apiext.api.controllers.image_imports
      operationId: import_image_config
      x-nextlinux-authz-action: importImage
      summary: "Import a docker or OCI image config to associate with the image"
      produces:
        - "application/json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
        - in: body
          name: contents
          required: true
          schema:
            type: object
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportContentResponse"
        500:
          description: "Internal Error"
definitions:
...
  StatusResponse:
    type: object
    description: System status response
    properties:
      available:
        type: boolean
      busy:
        type: boolean
      up:
        type: boolean
      message:
        type: string
      version:
        type: string
      db_version:
        type: string
      detail:
        type: object
...
  ApiErrorResponse:
    description: "Generic HTTP API error response"
    type: object
    properties:
      code:
        type: integer
        format: int32
      error_type:
        type: "string"
      message:
        type: "string"
      detail:
        type: object
        description: Details structure for additional information about the error if available. Content and structure will be error specific.
...
  ImageImportOperation:
    type: object
    description: An import record, creating a unique identifier for referencing the operation as well as its state
    properties:
      uuid:
        type: string
      status:
        type: string
        enum:
          - pending
          - queued
          - processing
          - complete
          - failed
          - expired
      expires_at:
        type: string
        format: date-time
      created_at:
        type: string
        format: date-time
  ImageImportContentResponse:
    type: object
    properties:
      digest:
        type: string
      created_at:
        type: string
        format: date-time
...
  ImageImports:
    type: array
    items:
      $ref: "#/definitions/ImageImportOperation"
  ImportContentDigestList:
    type: array
    items:
      type: string
      description: String digest of an uploaded content
  ImagePackageManifest:
    required:
      - artifacts
      - source
      - distro
    properties:
      artifacts:
        items:
          "$ref": "#/definitions/ImportPackage"
        type: array
      source:
        "$ref": "#/definitions/ImportSource"
      distro:
        "$ref": "#/definitions/ImportDistribution"
      descriptor:
        "$ref": "#/definitions/ImportDescriptor"
      schema:
        "$ref": "#/definitions/ImportSchema"
      artifactRelationships:
        items:
          "$ref": "#/definitions/ImportPackageRelationship"
        type: array
    additionalProperties: true
    type: object
  ImportDescriptor:
    required:
      - name
      - version
    properties:
      name:
        type: string
      version:
        type: string
    additionalProperties: true
    type: object
  ImportDistribution:
    required:
      - name
      - version
      - idLike
    properties:
      name:
        type: string
      version:
        type: string
      idLike:
        type: string
    additionalProperties: true
    type: object
  ImportPackageLocation:
    required:
      - path
    properties:
      path:
        type: string
      layerID:
        type: string
    additionalProperties: true
    type: object
  ImportPackage:
    required:
      - name
      - version
      - type
      - locations
      - licenses
      - language
      - cpes
      - metadataType
    properties:
      id:
        type: string
      name:
        type: string
      version:
        type: string
      type:
        type: string
      foundBy:
        type: string
      locations:
        items:
          "$schema": http://json-schema.org/draft-04/schema#
          "$ref": "#/definitions/ImportPackageLocation"
        type: array
      licenses:
        items:
          type: string
        type: array
      language:
        type: string
      cpes:
        items:
          type: string
        type: array
      purl:
        type: string
      metadataType:
        type: string
      metadata:
        type: object
    additionalProperties: true
    type: object
  ImportSchema:
    required:
      - version
      - url
    properties:
      version:
        type: string
      url:
        type: string
    additionalProperties: true
    type: object
  ImportSource:
    required:
      - type
      - target
    properties:
      type:
        type: string
      target:
        additionalProperties: true
    additionalProperties: true
    type: object
  ImportPackageRelationship:
    required:
      - parent
      - child
      - type
    properties:
      parent:
        type: string
      child:
        type: string
      type:
        type: string
      metadata:
        additionalProperties: true
        type: object
    additionalProperties: true
    type: object
...